+++ src/main/java/com/complete/todayspace/domain/chat/repository/ChatMessageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface ChatRepository extends MongoRepository<ChatMessage, String> {
+public interface ChatMessageRepository extends MongoRepository<ChatMessage, String> {
 
     List<ChatMessage> findByRoomIdOrderBySendDateAsc(String roomId);
