+++ src/main/java/com/complete/todayspace/global/aop/TimeTrackerAop.java
@@ -13,7 +13,7 @@ public class TimeTrackerAop {
 
     private static final Logger logger = LoggerFactory.getLogger(TimeTrackerAop.class);
 
-    @Around("execution(* com.complete.todayspace.domain.chat.service.ChatService.getChatRoom(..)) || "
+    @Around("execution(* com.complete.todayspace.domain.chat.service.ChatRoomService.getChatRoom(..)) || "
             + "execution(* com.complete.todayspace.domain.product.service.ProductService.getResponseDto(..)) || "
             + "execution(* com.complete.todayspace.domain.product.service.ProductService.getMyProductList(..))")
     public Object checkTime(ProceedingJoinPoint joinPoint) throws Throwable {
